Adding second branch for loopback, cleanup

+++ python/misc/widebuf.py
@@ -20,8 +20,8 @@ class WideBuf:
             if offset >= dst.size:
                 break
 
-    def decompose(self, src, numSamples, bufferNum, buffersCount, samplesPerFLit):
-        buffer = np.empty(numSamples, dtype=np.uint16)
+    def decompose(self, src, numSamples, bufferNum, buffersCount, samplesPerFLit, dtype=np.uint16):
+        buffer = np.empty(numSamples, dtype=dtype)
         offset = bufferNum * samplesPerFLit
         for i in range(0, buffer.size, samplesPerFLit):
